requirements.txt
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
PyYAML>=6.0.1
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
